clean up; remove unused refactoring
