--- src/store/Filings/index.ts
@@ -1,12 +1,0 @@
-import actions from './actions'
-import getters from './getters'
-import mutations from './mutations'
-import { state } from './state'
-
-/** The store configuration module. */
-export default {
-  actions,
-  getters,
-  mutations,
-  state
-}
