--- src/store/Filings/state.ts
@@ -1,11 +1,0 @@
-import { RestorationIF } from '@/interfaces'
-
-export const state: any = {
-  businessInfo: {},
-  addresses: {},
-  nameTranslations: [],
-  orgPersons: [],
-  authInfo: {},
-  stateFilingData: {},
-  restorationFiling: {} as RestorationIF
-}
